Replace LCTRL <-> LGUI
I use LCTRL *very* often in the iTerm. So it's very conve-
nient to be assigned on the <left thumb> (to the left of
the SPACE key). And almost never use LGUI in the iTerm.

But outside the terminal, it's the opposite, I almost never
use LCTRL, while using LGUI a lot.

Looks like I can remap modifiers in iTerm2, so I'll try to
use the same modifier everywhere, but in GUI apps it will
be actual LGUI key, while in the terminal, it automatically
be switched to the LCTRL.

Seems like a good idea, let's try.